[intl-tel-input] Hide types that are not imported in modules.
