+++ types/intl-tel-input/test/module-tests.ts
@@ -2,3 +2,8 @@ import intlTelInput = require('intl-tel-input');
 
 // $ExpectType (node: Element, options?: Options) => Plugin
 intlTelInput;
+
+// $ExpectError
+const options: IntlTelInput.Options = {
+    allowDropdown: true
+};
